Remote modal

+++ client/src/modules/packageinstaller.js
@@ -107,6 +107,18 @@ export default class PackageInstaller {
      * @param {String} remoteLocation Remote resource location
      */
     static async installRemotePackage(remoteLocation) {
+        let outputPath = null;
+
+        try {
+
+            const modalResult = await Modals.remoteInstallModal(remoteLocation).promise;
+            console.log(modalResult);
+            await this.dragAndDropHandler(modalResult);
+        } catch (err) {
+            throw err;
+        }
+
+        return;
         try {
             const { hostname } = Object.assign(document.createElement('a'), { href: remoteLocation });
             if (hostname !== 'api.github.com' && hostname !== 'secretbdapi') throw 'Invalid host!';
@@ -121,18 +133,78 @@ export default class PackageInstaller {
             };
 
             const response = await request.get(options);
-            const outputPath = path.join(Globals.getPath('tmp'), Security.hash('sha256', response, 'hex'));
+            outputPath = path.join(Globals.getPath('tmp'), Security.hash('sha256', response, 'hex'));
             fs.writeFileSync(outputPath, response);
-            console.log('response', response);
-            console.log('output', outputPath);
 
-            await this.dragAndDropHandler(outputPath);
+            const config = JSON.parse(asar.extractFile(outputPath, 'config.json').toString());
+            const { info, main } = config;
+
+            let icon = null;
+            if (info.icon && info.icon_type) {
+                const extractIcon = asar.extractFile(outputPath, info.icon);
+                icon = `data:${info.icon_type};base64,${Utils.arrayBufferToBase64(extractIcon)}`;
+            }
+
+            const isPlugin = info.type && info.type === 'plugin' || main.endsWith('.js');
+
+            // Show install modal
+            const modalResult = await Modals.installModal(isPlugin ? 'plugin' : 'theme', config, filePath, icon).promise;
+
+        } catch (err) {
+            throw err;
+        } finally {
+            if (!outputPath) return;
             rimraf(outputPath, err => {
                 if (err) console.log(err);
             });
+        }
+    }
+
+    static async downloadRemotePackage(remoteLocation) {
+        let outputPath = null;
+
+        try {
+            const { hostname } = Object.assign(document.createElement('a'), { href: remoteLocation });
+            if (hostname !== 'api.github.com' && hostname !== 'secretbdapi') throw 'Invalid host!';
+
+            const options = {
+                uri: remoteLocation,
+                encoding: null,
+                headers: {
+                    'User-Agent': 'BetterDiscordClient',
+                    'Accept': 'application/octet-stream'
+                }
+            };
+
+            const response = await request.get(options);
+            outputPath = path.join(Globals.getPath('tmp'), Security.hash('sha256', response, 'hex'));
+            fs.writeFileSync(outputPath, response);
+
+            const config = JSON.parse(asar.extractFile(outputPath, 'config.json').toString());
+            const { info, main } = config;
+
+            let icon = null;
+            if (info.icon && info.icon_type) {
+                const extractIcon = asar.extractFile(outputPath, info.icon);
+                icon = `data:${info.icon_type};base64,${Utils.arrayBufferToBase64(extractIcon)}`;
+            }
+
+            const isPlugin = info.type && info.type === 'plugin' || main.endsWith('.js');
+
+            return {
+                response,
+                outputPath,
+                config,
+                icon,
+                isPlugin
+            };
+
         } catch (err) {
             throw err;
+        } finally {
+
         }
+
     }
 
     /**
